Remove some unused storage callbacks
A lot of these functions are afaict just defined on the interface,
implemented in Signal-Android, and then never used anywhere at all. A
few are used within the Signal-Android client, but not in
libsignal-protocol-{java,rust}. Remove anything that we ourselves do
not need implemented, because every additional function on these
traits adds a few dozen lines in the JNI interop layer.

This doesn't negatively affect Signal-Android since we will retain the
various functions on the Java interface specs. They'll simply never be
invoked, which is mostly the status quo anyway with
libsignal-protocol-java.